backend/requirements.txt
absl-py==2.1.0
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2==2.9.10
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
